+++ src/sessions/softphone-session-handler.ts
@@ -338,7 +338,7 @@ export default class SoftphoneSessionHandler extends BaseSessionHandler {
       return;
     }
 
-    const participantsForUser = update.participants.filter(p => p.userId === this.sdk._personDetails.id);
+    const participantsForUser = update.participants.filter(p => p.userId === this.sdk._personDetails.id).reverse();
     let participant: IConversationParticipantFromEvent;
 
     if (!participantsForUser.length) {
@@ -354,11 +354,6 @@ export default class SoftphoneSessionHandler extends BaseSessionHandler {
     /* find user participant with desired call state */
     if (!participant && state) {
       participant = participantsForUser.filter(p => p.calls.find(c => c.state === state))[0];
-    }
-
-    /* find the most recent participant that is not already terminated or disconnected */
-    if (!participant) {
-      participant = participantsForUser.filter(p => p.calls.some(c => !this.isEndedState(c)))[0];
     }
 
     /* find user participant with a call */
@@ -523,14 +518,6 @@ export default class SoftphoneSessionHandler extends BaseSessionHandler {
   }
 
   async rejectPendingSession (pendingSession: IPendingSession): Promise<any> {
-    if (!this.hasActiveSession()) {
-      return super.rejectPendingSession(pendingSession);
-    }
-    this.log('info', 'rejecting pending session with an active persistent connection', {
-      sessionId: pendingSession.id,
-      conversationId: pendingSession.conversationId
-    });
-
     let participant = this.getUserParticipantFromConversationEvent(
       this.conversations[pendingSession.conversationId]?.conversationUpdate
     );
@@ -539,8 +526,19 @@ export default class SoftphoneSessionHandler extends BaseSessionHandler {
       participant = await this.fetchUserParticipantFromConversationId(pendingSession.conversationId);
     }
 
+    if (participant.purpose === 'user') {
+      return this._rejectUcCall(pendingSession.conversationId, participant.id);
+    }
+
     return this.patchPhoneCall(pendingSession.conversationId, participant.id, {
       state: CommunicationStates.disconnected
+    });
+  }
+
+  _rejectUcCall (conversationId: string, participantId: string): Promise<any> {
+    return requestApi.call(this.sdk, `/conversations/calls/${conversationId}/participants/${participantId}/replace`, {
+      method: 'post',
+      data: JSON.stringify({ voicemail: true })
     });
   }